# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: 群想 AI 群聊 API
  version: 1.0.0
  description: |
    以 backend/src 真实路由为准的接口契约。覆盖认证、短信、群组、消息、文件、
    TTS、社交互动、记忆、Agent、互动分析、监控、画像、人设与用户 API 配置等域。
    部分复杂查询端点（搜索、互动分析、监控指标等）的响应体为宽松对象，
    字段以实现为准，后续逐步收紧。
servers:
  - url: http://localhost:3002
    description: 本地后端
  - url: https://{YOUR_BACKEND_HOST}
    description: 生产后端（占位符，替换为真实 Render/自有域名）
    variables:
      YOUR_BACKEND_HOST:
        default: your-backend-host.example.com
...
security:
  - cookieAuth: []
  - csrfHeader: []
paths:
...
  /api/auth/login-phone:
    post:
      tags: [Auth]
      operationId: loginWithPhone
      summary: 手机号 + 密码登录（生产认证链路）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneLoginInput'
      responses:
        '200':
          description: 登录成功，下发 session_token Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 手机号或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/register-sms:
    post:
      tags: [Auth]
      operationId: registerBySms
      summary: 短信验证码注册（生产注册链路）
      description: |
        需先调用 POST /api/sms/send 获取验证码。密码必须包含大小写字母与数字。
        注册成功自动建立会话并初始化用户数据库。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsRegisterInput'
      responses:
        '201':
          description: 注册成功，下发 session_token Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 该手机号已注册
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: 短信服务未配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: devRegisterUser
      summary: '[dev-only] 用户名密码注册'
      x-dev-only: true
      deprecated: false
      description: |
        ⚠️ 仅在 NODE_ENV != production 时注册该路由，生产环境 404。
        仅供本地开发与自动化测试使用；生产注册必须走 /api/auth/register-sms。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevRegisterInput'
      responses:
        '201':
          description: 注册成功，下发 session_token Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: 获取当前登录用户
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/AuthUser'
        '401':
          description: 未登录（user:null + requiresAuth:true）
          content:
            application/json:
              schema:
                type: object
                required: [user, requiresAuth]
                properties:
                  user:
                    nullable: true
                  requiresAuth:
                    type: boolean

  # --------------------------------------------------------------------- Sms
...
  /api/ai/models/{modelId}:
    get:
      tags: [Ai]
      operationId: getAiModelStats
      summary: 获取单个模型统计
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      responses:
        '200':
          description: 模型统计
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
  /api/ai/models/{modelId}/enabled:
    put:
      tags: [Ai]
      operationId: setModelEnabled
      summary: 启用/禁用模型
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          description: 操作结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
  /api/ai/models/{modelId}/config:
    put:
      tags: [Ai]
      operationId: updateModelConfig
      summary: 更新模型运行配置
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 操作结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/ai/models/{modelId}/reset-circuit-breaker:
    post:
      tags: [Ai]
      operationId: resetModelCircuitBreaker
      summary: 重置模型熔断器
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      responses:
        '200':
          description: 重置结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: 登录端点下发的 httpOnly 会话 Cookie
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
      description: |
        所有非安全方法（POST/PUT/PATCH/DELETE）必须携带，
        令牌通过 GET /api/csrf-token 获取，且须与 Cookie 中的
        CSRF 值一致（double-submit 模式）
  parameters:
...
    ModelIdParam:
      in: path
      name: modelId
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id: { type: string }
        username: { type: string }
        nickname: { type: string }
        phone: { type: string }
    AuthResponse:
      type: object
      required: [success, user]
      properties:
        success:
          type: boolean
          const: true
        user:
          $ref: '#/components/schemas/AuthUser'
...
    PhoneLoginInput:
      type: object
      required: [phone, password]
      additionalProperties: false
      properties:
        phone:
          type: string
          pattern: '^1[3-9][0-9]{9}$'
        password: { type: string }
...
    SmsRegisterInput:
      type: object
      required: [phone, password, code]
      additionalProperties: false
      properties:
        phone:
          type: string
          pattern: '^1[3-9][0-9]{9}$'
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: 必须包含大写字母、小写字母和数字
        code:
          type: string
          pattern: '^[0-9]{4,8}$'
        nickname:
          type: string
          maxLength: 50
    DevRegisterInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          pattern: '^[a-zA-Z0-9_-]{3,64}$'
        password:
          type: string
          minLength: 8
          maxLength: 128
        nickname: { type: string }
...